refactor: `full_context_label` リネーム
